Convert page to component
